Update spotify rest type names
It seems like the main use of this typename is to build the cache key
for apollo cache (e.g. a song result with the id 123 will be saved in
the cache as Song:123). Songs of listens will also be cached as
Song:123. But if the rest typename is Song! then Song!:123 will be
saved as a separate entry from Song:123 in the apollo cache.