Minor refactor of progress bar
I was hoping this would fix the async update to an unmounted component
warning, but it seems like the interval is just firing too rapidly.

Either way, I think it looks better (?).

>>>
Warning: Can't perform a React state update on an unmounted
component. This is a no-op, but it indicates a memory leak in your
application. To fix, cancel all subscriptions and asynchronous tasks
in a useEffect cleanup function.