Route Protection (#8)
* Use one CALIBRATE event instead of CALIBRATED_TO_*

* Add more realistic Given step for day/after sunset/before sunrise

* Always restore clock before setting a new one

* Redirect to listens page from question and submit page at night

* Move route protection to its own feature

* Move step for going back/forward to common

* Clear interval when progress bar unmounts

* Add lastSunrise to the sundial info stored in apollo cache

* Add useSubmittedAfterLastSunrise

* Add scenario: I try to visit the submit page after submitting

* Fix typo

* Add scenarios for a past submit being saved in local storage

* Add server stubs to RouteProtection feature to stop real requests

* Minor refactor of progress bar

I was hoping this would fix the async update to an unmounted component
warning, but it seems like the interval is just firing too rapidly.

Either way, I think it looks better (?).

>>>
Warning: Can't perform a React state update on an unmounted
component. This is a no-op, but it indicates a memory leak in your
application. To fix, cancel all subscriptions and asynchronous tasks
in a useEffect cleanup function.

* Fix typo

* Improve step language

* Improve scenario name

* Use clearer step names for waiting for submit to complete

* Don't record cypress to dashboard in CI (see #9)