First idea for caching sunlight windows
I thought I had access to the updated cache on an afterware link. This
plan was to add an afterware link: if the operation was
SunlightWindows query, then just pluck out the newly cached sunlight
windows and persist them to local storage.

But: cache _isn't_ updated at the point afterware can look at the
cache.